requirements.txt
click==8.0.1
colorama==0.4.4
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
greenlet==1.1.1
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.0
PyMySQL @ file:///home/conda/feedstock_root/build_artifacts/pymysql_1610208984992/work
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
